fix: 엔티티 컬럼 변경 (좋아요→평점)

+++ src/main/java/net/cafree/domain/feed/dto/request/FeedUpdateRequest.java
@@ -5,7 +5,7 @@ import lombok.Builder;
 import java.util.List;
 
 public record FeedUpdateRequest(String contents,
-                                Integer likePoint,
+                                Double likePoint,
                                 List<String> tags,
                                 Long cafeId) {
